fix incorrect format string [touch: 10546]
This accounts for the gotcha with the new `VsprintFilter` where the full html argument will always be located at the index after the arguments submitted in the constructor (which in this case is 3).  So we want to make sure that our format excludes the full html render to prevent wonkiness.