add fixes for bugs exposed by unit tests [touch: 10546]
One critical thing I caught with this was the use of the static keyword to “cache” the message_template_group.  This wouldn’t have caused problems with EventSmart because its not used for controlling any message sending, however it does make things difficult for tests.

The static was overkill anyways because the message template group call would have returned the model cached object anyways when the same id is requested.