change theshold part of query [touch: 10953]
The change here is that instead of considering the first 24 hours part of the date query, we calculate beginning AFTER the first 24 hours.  Thus when the query is done, we’re actually getting results between 1 day after the time of the query and 1 day * the threshold setting.

This will produce more accurate results  than the original query (which would actualy seem to retrieve things one day earlier than the indicated threshold).