Bumping version to 1.0.6.rc.002. Previous Commit message: "Add 'post' event notifications. (#18)
* Add days after threshold constant;
Add post event and datetime message type constants;

* Add Post Event message type;

* Add Post Datetime message type;

* Add Post Event message validator;

* Add Post Datetime message validator;

* Add 'post' message types to allowed message types.

* Add the various message templates for 'post' events.
(May want to skip these for review)

* Add getMessageTypeMetaKey method.
This is used to determine if the 'threshold' value should be set for 'days before' or 'days after' meta key based on the current message type (upcoming or post).

* If the current message type is an 'upcoming' message type, output a subsection using META_KEY_DAYS_BEFORE_THRESHOLD as the key, otherwise it's a 'post' message so use the META_KEY_DAYS_AFTER_THRESHOLD;

Depending on which of the META_KEYS is set as a key on metabox, use that key to set the 'threshold';

Add 'post' event message types to 'getContentString' method (used to output the string used on the metabox, basically 'before' vs 'after');

* Allow the schedule form to load for 'post' message types;

* Add PostEventNotificationsCommand
(This is a copy of UpcomingEventNotificationsCommand using the correct message type constant)

* Copy UpcomingEventNotificationsCommandHandler and rename to PostEventNotificationsCommandHandler;

* Rename 'Upcoming' to 'Post';

* Use POST event message type for this command.

* Adjust query to use the post message type and DTT_EVT_end (post event messages are sent after the event has expired).

* itemsToExclude should be expired events here.

* Add PostNotificationsCommand
(Copy of UpcomingNotificationsCommand)

* Copy UpcomingNotificationsCommandHandler and rename to PostNotificationsCommandHandler.

* Update 'Upcoming' to 'Post'

* handle() uses a PostNotificationsCommand

* Update filters to use 'Post' in place of 'Upcoming';

* Pass the SchedulingSettings and context to getStartTimeForQuery();
Alter the startTime to time() - ( {theshold in days} + cron buffer ) so that the query is pulling expired events/datetimes.

* Set up getEndTimeForQuery to be (time() - {theshold in days}) + {cron buffer}

This gives around a 7 hour window for the cron to run and trigger the messages.

* Pass the scheduleSettings and context to getStartTimeForQuery();

* Add PostDatetimeNotificationsCommand
(Copy of UpcomingDatetimeNotificationsCommand)

* Add PostDatetimeNotificationsCommandHandler.
(This is a copy of the upcoming version which has then had the query altered to work for POST Datetimes)

* Items to exclude should be 'expired'.

* Add Post command hanlders;
Add Post event message types;
Add command handler dependencies;

* Add post event checks to the cron schedule.

* Update Author and Since for any new classes I've added.

* Add @Since to new domain constants

* Use switch to determine if we should use DAYS_AFTER or DAYS_BEFORE threshold.

* Confirm we have a value for DAYS_BEFORE or DAYS_AFTER threshold, if not return false as we can't process anything.

* Initialize subsections array before using it.

* Set allowed_message_types to public.

* Catch if either DAYS_BEFORE or DAYS_AFTER threshold has been set to 0 or not set at all. A value other than 0 for either on of those thresholds is acceptable.

* Take 2! (Only committed a single file on previous commit)

Anywhere we are checking for all message types use EED_Automated_Upcoming_Event_Notification_Messages::allowed_message_types()

* Allow post datetime message types to use EE_Specific_Datetime_Shortcodes

* Set EE_Specific_Datetime_Shortcodes to allow for the Post_Datetime message type

* Fix conditional for datetime specific shortcodes.

Return nothing if we don't have either a EE_Automate_Upcoming_Datetime_message_type or EE_Automate_Post_Datetime_message_type OR no EE_Datetime object.

* Use DTT_EVT_end for the query to pull datetimes.

* write_log() call no longer needed.
It was used to confirm the query used to pull events to notify.

* Set the query StartTime to be (time() - current threshold in days) - a 'buffer' which by default will be 7 days.
Set the query EndTime to be time() - current threshold in days.
(Meaning post messages can trigger up to 7 days 'over' the threshold and will then be ignored)

* No need for a cron_frequency_buffer property;
Remove setCronFrequencyBuffer() method and just use getCronFrequencyBuffer() directly (its already filtered to allow for changes to the buffer);
Set the default cron buffer to 7 days;

* Update filter to be specific to PostNotificationsCommandHandler;

* Update comment regarding sending messages per datetime."